requirements_master.txt
absl-py==0.8.0
astor==0.8.0
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.9.11
chardet==3.0.4
cloudpickle==1.2.2
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
future==0.17.1
gast==0.2.2
gin-config==0.1.3
google-pasta==0.1.7
grpcio==1.23.0
gym==0.10.11
h5py==2.10.0
idna==2.8
imageio==2.4.0
imageio-ffmpeg==0.3.0
importlib-metadata==0.23
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.15.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.3
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
more-itertools==7.2.0
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.1
numpy==1.17.0
opt-einsum==3.0.1
packaging==19.2
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
Pillow==6.1.0
pluggy==0.13.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
py==1.8.0
pyglet==1.3.2
Pygments==2.4.2
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.0.1
python-dateutil==2.8.0
pywin32==225
pywinpty==0.5.5
pyzmq==18.1.0
qtconsole==4.5.5
requests==2.22.0
scipy==1.3.1
Send2Trash==1.5.0
six==1.12.0
tb-nightly==1.15.0a20190911
tensorflow-probability==0.7.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-agents-nightly==0.2.0.dev20190907
tf-estimator-nightly==1.14.0.dev2019092001
tf-nightly==1.15.0.dev20190821
tfp-nightly==0.9.0.dev20190920
tornado==6.0.3
traitlets==4.3.2
urllib3==1.25.5
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==0.6.0
